Match works on more types, and can map to values
The types of the matching functions are now more complex, because they
can work on MediaTypes, ByteStrings, and Quality marked values. The
standard approach to parsing and then matching values remains the same,
however. The match function can work without quality values on the
server side, and can directly map to a value (unimplemented).