Change API to always parse headers, remove quality
The previous API required that the client input was a list of quality
tagged values, but only provided the ability to tag media types during
parsing. Now the quality tagging is completely gone from the external
API, and all of the match* and map* functions automatically parse their
input based on an additional method in the type class, which has been
renamed from Match to Accept to reflect its expanded focus.

Tests for Content-Type matching have also been added. This change bumps
the version number to 0.2.