Add Accept-Language header data and parsing
As language negotiation also involves a degree of matching more
complicated than just equality checking, it is now represented by a
Language datatype with an Accept instance that handles matching and
specificity. Parsing and matching of quality annotated languages is
handled as before, due to the generic type signature of the related
functions.