Add support for Travis builds
Travis will now build and test the package as Haskell. The current
version of cabal on Travis has a different interface to the latest
testing inferface, so the script installs the latest version in a
sandbox and then uses that to configure and test.

The build status has been added to the readme.