Allow whitespace in MediaType parseAccept
Space and horizontal tab characters are allowed around the semicolon
character when parsing a media type in an HTTP header.

This patch allows this by:

- Repurposing the previously unused trimBS to trim both spaces and tabs
- After splitting a media type on its semicolons, trimming each part

The patch only deals with media types, not any of the other Accept-X
headers, but the trimBS function should come in handy for these as well.
A padding generator for testing is included as well.