Bump dependencies and modify test suite
The new dependencies permit the stable release of utf8-string, and
limits the lower bound on Cabal in the testing to avoid
incompatibilities with the detailed testing interface. Travis should now
test GHC 7.6 and 7.8.

The test options have also been updated to not generate HPC coverage
information (as this is left up to Cabal), and error out on warnings.