Replace UTF-8 interpretation of headers with Char8
The standard for header value encoding is just ISO 8859-1 (or more
practically just ASCII), not any Unicode representation (ignoring
encoded-word syntax). As such, the project's utf8-string dependency has
been removed and the Char8 module in bytestring is used instead. This
change permits the use of Char instead of Word8 on the individual bytes
in the strings, removing the hard-coded mappings in the Utils module and
replacing them with more recognisable character literals.

Some of the valid reg-name character bytes were wrong as well, and this
has been corrected to the right set of characters.