Perform case insensitive matching on strings
The matching behaviour of all of the header values should be case
insensitive. The internals of the MediaType and Language definitions are
now stored as case insensitive bytestrings (compatible with the header
standard of ISO 8859-1), forcing the accompanying functions to deal with
a folded-case form of the input. The Accept instance for ByteString
matches through equality of case-insensitive forms of its arguments. Now
media types, languages, and general bytestrings will match in a case
insensitive way, and media types and languages implement Eq with no
respect to case either.

The interface into creating and parsing the various Accept types remains
unchanged, and all input bytestrings are automatically converted to
case-insensitive form when stored. The primary difference in the
interface is the fields of the Accept types, which now produce case
insensitive bytestrings instead of the regular form.

Note that no specific tests have been added, as the signature change
requires that the tests naturally consider the case insensitive
requirement. Many of the tests now produce their input from a generated
case insensitive value.