Allow whitespace in quality parameters
The parseQuality function now handles whitespace around the semicolon
separating the quality parameter from the rest of the header. The
parsing now also explicitly handles extension parameters on the Accept
header, though it does not try to parse them as parameters and just
throws them away. All other Accept-X header values are expected to have
exactly one semicolon if they are equipped with a quality parameter.