Use case-insensitive equality in genDiffByteString
Generating a different bytestring than the given one used standard byte
equality, which meant the generator might produce a bytestring that was
equal under case-insensitive comparison, and thus cause the test that
required that different bytestrings not match to fail. Fortunately this
very rare case was caught by Travis, and the definition of the generator
now defers to the case-insensitive generator, wrapping and unwrapping
the input and output.