Relax Cabal lower bound, simplify Travis install
The lower bound on Cabal was raised to 1.19 to avoid conflicting with a
change from 1.18, which made testing more complicated (as 1.18 is the
Cabal included in GHC 7.8). It turns out the conflict can be avoided by
reducing the lower bound to 1.16, rather than raising it to 1.19, which
allows the Travis configuration to be simplified back to its defaults.