Fix type errors and import warnings in GHC 7.10
The BBP created an ambiguity between the new Foldable form of notElem
and a string literal under OverloadedStrings. This has been clarified by
translating the string literal into an actual list of characters.

The addition of a number of important combinators to the prelude raises
a number of redundant import warnings throughout the project and tests.
These have been resolved with the CPP, ensuring the project will still
build with older versions of base. The excluded TupleSections extension
has been added to the other-extensions for the test-suite in the cabal
file.