Switch cabal-test-quickcheck for test-framework
Because of the changes to the detailed-0.9 testing system in Cabal
between versions 1.18 and 1.20, cabal-test-quickcheck can fail to work
if the version of Cabal and the version of cabal-install differ across
that boundary. This change switches to using the test-framework library
with the exitcode-stdio-1.0 testing system instead. The change is mostly
imports, as the testing interface is the same.