Ensure parse failures do not raise exceptions
The breakChar function could raise an exception if the character was not
in the string. Now it returns a Maybe result and parse errors that
previously raise exceptions will now return Nothing. Several new test
cases have been added to ensure that bad inputs to parseAccept do not
raise exceptions.