Allow wildcards in MediaType creation helper
The (//) MediaType creation operator ensures that the two strings you
give it are valid reg-names, but as a result didn't support
wildcards (as * is not a valid reg-name character). Now the operator
ensures that either both arguments are wildcards, the main-type is a
valid reg-name and the sub-type is a wildcard, or both types are valid
reg-names.