Update tests to produce counter-examples on fail
The existing tests mostly resulted in boolean properties, which could
not be resolved into counter-examples when they failed. Now the tests on
equality use the QuickCheck operators (===) and (.&&.) to build
informative properties, and uses of the isNothing function have been
replaced with (=== Nothing). Tests of the predicate functions (like
matches) remain as boolean tests.