Correct faulty tests for match* and map* functions
The Nothing test for the match* functions failed on particular inputs,
which revealed that it was filtering in the wrong direction. Now they
don't even attempt to build different media types, and filter in the
opposite order. The tests for matchContent and mapContent weren't
actually testing those functions, as they were using matchAccept and
mapAccept. Some of the tests which were not relevant for the actual
functions have been removed, and the rest have been updated to reflect
the intended behaviour of the *Content functions.