Update project information, raise QuickCheck bound
This change updates the maintainer email, the copyright range, and the
gitignore file. The Travis test settings have also been updated to
include newer versions of GHC. A generator for the tests which did not
correctly enforce the desired property that the two components it
generated did not match one another has also been fixed.

The upper bound of QuickCheck has been extended to include the latest
version, which bumps the version number of this package.