Fix #42: return mapped value immediately
The old behaviour for a map lookup was to search the keyset for a match
and then lookup the resulting value as a second operation on the map
afterwards. This lookup was using matching rather than equality, which
would work the same as equality for Accept because any earlier matching
key would have been the matched key, but for Content the matching is
inverted, which means that more specific keys earlier than the matched
key would not be the match, but would match in the subequent lookup.

This change removes the behaviour to transform the map to the keyset and
performs the lookup on the keys in the pairs directly, so that once a
match is found we can immediately return it. The same behaviour has been
added to the quality mapping, which would have worked correctly but was
less efficient.